storage: data_bin value is FLOAT, not DOUBLE. This fixes #5648.

+++ sql/mysql_schema.sql
@@ -799,7 +799,7 @@ CREATE TABLE data_bin (
   id_metric int NOT NULL,
   ctime int NOT NULL,
   status enum('0', '1', '2', '3', '4') NOT NULL default '3',
-  value double default NULL,
+  value float default NULL,
 
   FOREIGN KEY (id_metric) REFERENCES metrics (metric_id)
     ON DELETE CASCADE,
